// cms/src/payload-types.ts
export interface Product {
  id: string;
  name: string;
  slug?: string | null;
  category: 'Vape' | 'Hardware' | 'THCA' | 'E-Liquid' | 'Glass' | 'Accessories' | 'Featured';
  price: number;
  shortDescription?: string | null;
  featured?: boolean | null;
  active?: boolean | null;
  updatedAt: string;
  createdAt: string;
}